# src/Data_Module/openapi.yaml
openapi: 3.1.0
info:
  title: Data Module API
  description: Intelligent Logistics - Data Module (Source of Truth)
  version: 1.0.0
servers:
  - url: http://10.255.32.82:8080/api/v1
    description: Production server
paths:
...
  /arrivals:
    get:
      tags: [Arrivals]
      summary: List Arrivals
      description: Lists appointments with optional filters. Used by operator frontend to list daily arrivals.
      operationId: list_arrivals
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: gate_id
          in: query
          schema:
            type: integer
            nullable: true
        - name: shift_gate_id
          in: query
          schema:
            type: integer
            nullable: true
        - name: shift_type
          in: query
          schema:
            type: string
            nullable: true
        - name: shift_date
          in: query
          schema:
            type: string
            format: date
            nullable: true
        - name: status
          in: query
          schema:
            type: string
            nullable: true
        - name: scheduled_date
          in: query
          schema:
            type: string
            format: date
            nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
  /arrivals/stats:
    get:
      tags: [Arrivals]
      summary: Get Arrivals Stats
      description: Arrival statistics by status. Used in operator dashboard.
      operationId: get_arrivals_stats
      parameters:
        - name: gate_id
          in: query
          schema:
            type: integer
            nullable: true
        - name: target_date
          in: query
          schema:
            type: string
            format: date
            nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
  /arrivals/next/{gate_id}:
    get:
      tags: [Arrivals]
      summary: Get Upcoming Arrivals
      description: Next scheduled arrivals for a gate. Used in operator's sidebar panel.
      operationId: get_upcoming_arrivals
      parameters:
        - name: gate_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
  /arrivals/{appointment_id}:
    get:
      tags: [Arrivals]
      summary: Get Arrival
      description: Gets details of a specific appointment.
      operationId: get_arrival
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
  /arrivals/pin/{arrival_id}:
    get:
      tags: [Arrivals]
      summary: Get Arrival By Pin Code
      description: Gets appointment by arrival_id/PIN. Used by driver to check their appointment.
      operationId: get_arrival_by_pin
      parameters:
        - name: arrival_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
...
  /arrivals/{appointment_id}/status:
    patch:
      tags: [Arrivals]
      summary: Update Status
      description: Updates appointment status. Used by operator or system for manual updates.
      operationId: update_status
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
...
  /arrivals/{appointment_id}/visit:
    post:
      tags: [Arrivals]
      summary: Create Visit
      description: Creates a visit when truck arrives. Uses composite FK to Shift.
      operationId: create_visit
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisitRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
    patch:
      tags: [Arrivals]
      summary: Update Visit
      description: Updates visit status (e.g., to 'completed' when truck leaves).
      operationId: update_visit
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'

  # ==================== DECISIONS ====================
...
components:
  schemas:
    Appointment:
      type: object
      required: [booking_reference, driver_license, truck_license_plate, terminal_id, id, arrival_id]
      properties:
        id:
          type: integer
        arrival_id:
          type: string
        booking_reference:
          type: string
        driver_license:
          type: string
        truck_license_plate:
          type: string
        terminal_id:
          type: integer
        gate_in_id:
          type: integer
          nullable: true
        gate_out_id:
          type: integer
          nullable: true
        scheduled_start_time:
          type: string
          format: date-time
          nullable: true
        expected_duration:
          type: integer
          nullable: true
        status:
          $ref: '#/components/schemas/AppointmentStatusEnum'
        notes:
          type: string
          nullable: true
        booking:
          $ref: '#/components/schemas/Booking'
          nullable: true
        driver:
          $ref: '#/components/schemas/Driver'
          nullable: true
        truck:
          $ref: '#/components/schemas/Truck'
          nullable: true
        terminal:
          $ref: '#/components/schemas/Terminal'
          nullable: true
        gate_in:
          $ref: '#/components/schemas/Gate'
          nullable: true
        gate_out:
          $ref: '#/components/schemas/Gate'
          nullable: true
    AppointmentStatusEnum:
      type: string
      enum: [in_transit, canceled, delayed, completed]
    AppointmentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/AppointmentStatusEnum'
        notes:
          type: string
          nullable: true
    Visit:
      type: object
      required: [appointment_id, shift_gate_id, shift_type, shift_date]
      properties:
        appointment_id:
          type: integer
        shift_gate_id:
          type: integer
        shift_type:
          $ref: '#/components/schemas/ShiftTypeEnum'
        shift_date:
          type: string
          format: date
        entry_time:
          type: string
          format: date-time
          nullable: true
        out_time:
          type: string
          format: date-time
          nullable: true
        state:
          $ref: '#/components/schemas/DeliveryStatusEnum'
    VisitStatusUpdate:
      type: object
      required: [state]
      properties:
        state:
          $ref: '#/components/schemas/DeliveryStatusEnum'
        entry_time:
          type: string
          format: date-time
          nullable: true
        out_time:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
    CreateVisitRequest:
      type: object
      required: [shift_gate_id, shift_type, shift_date]
      properties:
        shift_gate_id:
          type: integer
        shift_type:
          type: string
          enum: [MORNING, AFTERNOON, NIGHT]
        shift_date:
          type: string
          format: date
    DeliveryStatusEnum:
      type: string
      enum: [unloading, completed]
    ShiftTypeEnum:
      type: string
      enum: ['06:00-14:00', '14:00-22:00', '22:00-06:00']
    Booking:
      type: object
      required: [reference]
      properties:
        reference:
          type: string
        direction:
          $ref: '#/components/schemas/DirectionEnum'
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
    DirectionEnum:
      type: string
      enum: [inbound, outbound]
    Driver:
      type: object
      required: [name, drivers_license]
      properties:
        drivers_license:
          type: string
        name:
          type: string
        company_nif:
          type: string
          nullable: true
        mobile_device_token:
          type: string
          nullable: true
        active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
          nullable: true
        company:
          $ref: '#/components/schemas/Company'
          nullable: true
...
    Truck:
      type: object
      required: [license_plate]
      properties:
        license_plate:
          type: string
        company_nif:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        company:
          $ref: '#/components/schemas/Company'
          nullable: true
    Company:
      type: object
      required: [name, nif]
      properties:
        nif:
          type: string
        name:
          type: string
        contact:
          type: string
          nullable: true
    Terminal:
      type: object
      required: [id]
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        hazmat_approved:
          type: boolean
          default: false
    Gate:
      type: object
      required: [label, id]
      properties:
        id:
          type: integer
        label:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
...